+++ packages/ssx/src/scripts/app.js
@@ -123,7 +123,6 @@ ${routes.join(",\n")}
 const module = await getRoute(page.pattern)()
 const type = module[page.moduleName]
 types[page.moduleName] = type
-const shouldHydrate = module.hydrate !== false
 
 const store = createStore({ types, entities, middlewares, systems, autoCreateEntities: true })
 
@@ -132,7 +131,7 @@ const root = document.getElementById("root")
 mount(store, (api) => {
   const { route } = api.getEntity("router")
   return api.render(route)
-}, root, { hydrate: shouldHydrate })
+}, root)
 `
 }