+++ packages/web/src/mount.js
@@ -10,15 +10,13 @@ import { html, render } from "lit-html"
  * @param {Store} store - The application state store.
  * @param {(api: Api) => TemplateResult | null} renderFn - The root render function.
  * @param {HTMLElement | DocumentFragment} element - The DOM element to mount the template to.
- * @param {Object} [options] - Mount options.
- * @param {boolean} [options.hydrate=true] - Whether to hydrate the existing DOM.
  * @returns {() => void} An unsubscribe function
  */
-export async function mount(store, renderFn, element, options = {}) {
+export async function mount(store, renderFn, element) {
   const api = { ...store._api }
   api.render = createRender(api)
 
-  const shouldHydrate = options.hydrate !== false && element.hasChildNodes()
+  const shouldHydrate = element.hasChildNodes()
 
   // If we need to hydrate, load the heavy lifting only now
   if (shouldHydrate) {