+++ benchmarks/dashboard/inglorious/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @benchmarks/dashboard-inglorious
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@5.0.2
+
 ## 0.1.7
 
 ### Patch Changes