+++ examples/apps/web-components-ssx/CHANGELOG.md
@@ -1,5 +1,13 @@
 # web-components-ssx
 
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/ssx@2.0.2
+  - @inglorious/web@5.0.2
+
 ## 0.1.8
 
 ### Patch Changes