+++ examples/apps/web-components/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web-components
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@5.0.2
+
 ## 0.0.6
 
 ### Patch Changes