+++ examples/apps/web-logo/CHANGELOG.md
@@ -1,5 +1,13 @@
 # lit-todomvc
 
+## 0.1.44
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@5.0.2
+  - @inglorious/logo@3.0.2
+
 ## 0.1.43
 
 ### Patch Changes