+++ examples/apps/web-todomvc-ace/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web-todomvc-ace
 
+## 0.1.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@5.0.2
+
 ## 0.1.18
 
 ### Patch Changes