+++ examples/apps/web-todomvc-jsx/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web-todomvc-jsx
 
+## 0.1.29
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@5.0.2
+
 ## 0.1.28
 
 ### Patch Changes