+++ packages/ag-grid/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inglorious/ag-grid
 
+## 4.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@5.0.2
+
 ## 4.0.1
 
 ### Patch Changes