+++ packages/ssx/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inglorious/ssx
 
+## 2.0.2
+
+### Patch Changes
+
+- Fix Markdown pages being exported as default
+- Updated dependencies
+  - @inglorious/web@5.0.2
+
 ## 2.0.1
 
 ### Patch Changes