+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inglorious/ui
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@5.0.2
+  - @inglorious/charts@5.0.2
+
 ## 2.0.1
 
 ### Patch Changes